cast min max values correctly

+++ tests/integration/query/simple/with_group_max_test.go
@@ -291,7 +291,7 @@ func TestQuerySimple_WithGroupByStringWithoutRenderedGroupAndChildEmptyFloatMax_
 						},
 						{
 							"Name": "Alice",
-							"_max": -math.MaxFloat64,
+							"_max": float64(-math.MaxFloat64),
 						},
 					},
 				},
